refactor: Add link to historical comments in validationTransferColumns
This commit adds a link to the historical comments page for each row in the validationTransferColumns component. The link is wrapped in a button and displayed next to the existing button. Clicking on the link will navigate the user to the historical comments page for the corresponding client. This enhancement improves the user experience by providing easy access to the historical comments related to each transfer.

Note: This commit message is based on the provided code changes and recent user commits.